Ported the other branch into this one

+++ badges/queries.py
@@ -1,16 +1,17 @@
-import json
 from base64 import b64decode
 from datetime import datetime
+from mimetypes import guess_extension
 
-from django.core import serializers
 from django.core.files.base import ContentFile
 from django.core.files.storage import default_storage
+from django.db.models import Q
 
 from locations.models import Location
-from .models import Badge, Status
+from users.models import PromoterUser, AppUser
+from .models import Badge, RedeemedBadges
 
 
-def create_badge(badge, promoter):
+def create_badge(badge, user_id):
     # Creating badge
     badge_created = Badge()
 
@@ -30,9 +31,20 @@ def create_badge(badge, promoter):
 
     if badge.get("image"):
         try:
+            # Accepted format: data:<mime_type>;base64,<encoding>
             img_format, img_str = badge.get("image").split(';base64,')
-            ext = img_format.split('/')[-1]
-            badge_created.image = ContentFile(b64decode(img_str), name=str(badge_created.uuid) + '.' + ext)
+            _, mime_type = img_format.split(':')
+
+            # If MIME type is not image
+            if mime_type.split("/")[0] != "image":
+                raise NotAValidImage()
+
+            # Decoding image from base64
+            decoded_img = b64decode(img_str)
+            # Getting extension from MIME type
+            file_extension = guess_extension(mime_type)
+            # Storing image
+            badge_created.image = ContentFile(decoded_img, name=str(badge_created.uuid) + '.' + file_extension)
         except Exception:
             raise NotAValidImage()
 
@@ -41,14 +53,33 @@ def create_badge(badge, promoter):
     except Location.DoesNotExist:
         raise NotAValidLocation()
 
-    badge_created.promoter = promoter
+    badge_created.promoter = PromoterUser.objects.get(user_id=user_id)
     badge_created.save()
 
     return badge_created
 
 
-def get_badge():
+def get_badges():
     return Badge.objects.all()
+
+
+def redeem_badges_by_location(location_uuid, user_id):
+    # Getting all badges that are associated with a location and are "up"
+    redeemable_badges = Badge.objects.filter(Q(location=location_uuid),
+                                             Q(start_date__lte=datetime.now()),
+                                             Q(end_date__isnull=-True) | Q(end_date__gte=datetime.now()))
+
+    # Getting app user that's redeeming the badge
+    apper = AppUser.objects.get(user_id=user_id)
+
+    # Linking the App User with the Badges
+    for redeemable_badge in redeemable_badges:
+        try:  # Checking if the user has already redeemed this badge (and if so do nothing)
+            RedeemedBadges.objects.get(app_user=apper, badge=redeemable_badge)
+        except RedeemedBadges.DoesNotExist:
+            RedeemedBadges(app_user=apper, badge=redeemable_badge).save()
+
+    return redeemable_badges
 
 
 def get_badge_by_uuid(badge_uuid):
@@ -71,8 +102,15 @@ def patch_badge_by_uuid(badge_uuid, badge):
         badge_update.name = badge.get('name')
     if badge.get('description'):
         badge_update.description = badge.get('description')
-    if badge.get('start_date'):
+    # If changing both dates
+    if badge.get('start_date') and not badge.get('end_date'):
         badge_update.start_date = badge.get('start_date')
+    # If changing only start
+    elif badge.get('start_date'):
+        if badge_update.end_date > badge.get('start_date'):
+            badge_update.start_date = badge.get('start_date')
+        else:
+            raise StartDateAfterEndDate()
     if badge.get('end_date'):
         if badge.get('end_date') > badge_update.start_date:  # End must be after the start
             badge_update.end_date = badge.get('end_date')
@@ -88,57 +126,29 @@ def patch_badge_by_uuid(badge_uuid, badge):
 
     if badge.get("image"):
         try:
+            # Accepted format: data:<mime_type>;base64,<encoding>
+            img_format, img_str = badge.get("image").split(';base64,')
+            _, mime_type = img_format.split(':')
+
+            # If MIME type is not image
+            if mime_type.split("/")[0] != "image":
+                raise NotAValidImage()
+
+            # Decoding image from base64
+            decoded_img = b64decode(img_str)
+
             # Deleting previous image from storage
             default_storage.delete(badge.image.path)
-            # Decoding and storing new image
-            img_format, img_str = badge.get("image").split(';base64,')
-            ext = img_format.split('/')[-1]
-            badge_update.image = ContentFile(b64decode(img_str), name=str(badge_update.uuid) + '.' + ext)
+            # Getting extension from MIME type
+            file_extension = guess_extension(mime_type)
+            # Storing image
+            badge_update.image = ContentFile(decoded_img, name=str(badge_update.uuid) + '.' + file_extension)
         except Exception:
             raise NotAValidImage()
 
     badge_update.save()
 
     return badge_update
-
-
-def encode_badge(badge):
-    return json.loads(serializers.serialize("json",
-                                            badge,
-                                            fields=[
-                                                'uuid', 'name',
-                                                'description', 'image',
-                                                'start_date', 'end_date',
-                                                'location', 'status']))
-
-
-def decode_badge(data, admin):
-    json_data = json.loads(data)
-    badge = {
-        'name': json_data.get("name"),
-        'description': json_data.get("description"),
-        'location': json_data.get("location"),
-        'image': json_data.get("image"),
-        'status': json_data.get("status") if admin else Status.PENDING  # Only admin can change status
-    }
-
-    # Setting start date
-    if "start_date" in json_data:
-        try:
-            date = datetime.fromisoformat(json_data.get("start_date"))
-            badge["start_date"] = date
-        except Exception:
-            raise NotAValidStartDate()
-
-    # Setting end date
-    if "end_date" in json_data:
-        try:
-            date = datetime.fromisoformat(json_data.get("end_date"))
-            badge["end_date"] = date
-        except Exception:
-            raise NotAValidEndDate()
-
-    return badge
 
 
 class NotAValidLocation(Exception):
@@ -149,17 +159,9 @@ class NotAValidImage(Exception):
     pass
 
 
-class NotAValidStartDate(Exception):
-    pass
-
-
-class NotAValidEndDate(Exception):
+class StartDateAfterEndDate(Exception):
     pass
 
 
 class EndDateNotAfterStartDate(Exception):
-    pass
-
-
-class ErrorUpdate(Exception):
     pass
